incorporated qemu_aarch64_test into qemu_test and removed
